feat(gui-v2): add `base` layout
