+++ packages/nc-gui-v2/components.d.ts
@@ -90,9 +90,9 @@ declare module '@vue/runtime-core' {
     MdiPlusIcon: typeof import('~icons/mdi/plus-icon')['default']
     MdiPlusOutline: typeof import('~icons/mdi/plus-outline')['default']
     MdiReload: typeof import('~icons/mdi/reload')['default']
-    MdiTableArrowRight: typeof import('~icons/mdi/table-arrow-right')['default']
     MdiShieldLockOutline: typeof import('~icons/mdi/shield-lock-outline')['default']
     MdiStore: typeof import('~icons/mdi/store')['default']
+    MdiTableArrowRight: typeof import('~icons/mdi/table-arrow-right')['default']
     MdiTableBorder: typeof import('~icons/mdi/table-border')['default']
     MdiTrashCan: typeof import('~icons/mdi/trash-can')['default']
     MdiXml: typeof import('~icons/mdi/xml')['default']
